Reorganized all bitly-related classes under same namespace
