+++ composer.json
@@ -14,10 +14,10 @@
     "require": {
         "php": "^7.4 || ^8.0",
         "ext-json": "*",
-        "laminas/laminas-servicemanager": "^3.4",
-        "lstrojny/functional-php": "^1.14",
+        "laminas/laminas-servicemanager": "^3.6.4",
+        "lstrojny/functional-php": "^1.15",
         "shlinkio/shlink-config": "^1.0",
-        "symfony/console": "^5.1"
+        "symfony/console": "^5.2"
     },
     "require-dev": {
         "guzzlehttp/guzzle": "^7.2",
